Don't delete icon_players and icon_duration during a cleanup.

+++ generate_pdf.py
@@ -25,7 +25,7 @@ args = parser.parse_args()
 
 if(args.clean_images):
     for f in os.listdir('./Images'):
-        if(exists(os.path.join('./Images', f))):
+        if(exists(os.path.join('./Images', f) and f != 'icon_players.png' and f != 'icon_duration.png')):
             os.remove(os.path.join('./Images', f))
     sys.exit()
